ob123-fixed issue between modal implementation and checkout process

+++ src/hooks/useSummaryButton.ts
@@ -29,7 +29,22 @@ const useSummaryButton = ({ path }: { path: string }) => {
       secondButtonLink: PATHS.CHECKOUT,
       onClickFirstButton: () => {},
       onClickSecondButton: () => {
-        handleDeleteOrder();
+        Swal.fire({
+          title: 'Are you sure?',
+          html: '<p>Are you sure you want to change your order type?</p>',
+          showConfirmButton: true,
+          showCancelButton: true,
+          confirmButtonColor: '#182715',
+          confirmButtonText: 'Yes, take me to change order page',
+          cancelButtonColor: '#b80f0a',
+          cancelButtonText: 'Cancel',
+          background: '#fff5e1',
+        }).then((result) => {
+          if (result.isConfirmed) {
+            navigate(PATHS.CHECKOUT);
+            handleDeleteOrder();
+          }
+        });
       },
       disabled: false, // Always enabled
     },
